+++ app/models/team/shared.rb
@@ -20,7 +20,7 @@ class Team::Shared < Team
       team = super(params)
       return team unless team.valid?
 
-      plaintext_team_password = Crypto::Symmetric::Aes256.random_key
+      plaintext_team_password = team.new_team_password
       team.add_user(creator, plaintext_team_password)
       unless team.private?
         User::Human.admins.each do |a|