+++ app/utils/encryptable_transfer.rb
@@ -3,18 +3,12 @@
 class EncryptableTransfer
 
   def transfer(encryptable, receiver, sender)
-    transfer_password = new_transfer_password
-    encryptable.encrypt(transfer_password)
+    encryption_algorithm = receiver_encryption_algorithm(receiver)
 
-    encryptable.name = encryptable_destination_name(encryptable, receiver)
+    transfer_password = encryption_algorithm.random_key
+    encryptable.encrypt(transfer_password, encryption_algorithm)
 
-    encryptable.update!(
-      folder: receiver.inbox_folder,
-      sender_id: sender.id,
-      encrypted_transfer_password:
-        Base64.encode64(encrypted_transfer_password(transfer_password, receiver))
-    )
-    encryptable
+    update_encryptable(encryptable, receiver, sender, transfer_password)
   end
 
   def receive(encryptable, private_key, personal_team_password)
@@ -31,6 +25,18 @@ class EncryptableTransfer
 
   private
 
+  def update_encryptable(encryptable, receiver, sender, transfer_password)
+    encryptable.name = encryptable_destination_name(encryptable, receiver)
+
+    encryptable.update!(
+      folder: receiver.inbox_folder,
+      sender_id: sender.id,
+      encrypted_transfer_password:
+        Base64.encode64(encrypted_transfer_password(transfer_password, receiver))
+    )
+    encryptable
+  end
+
   def encryptable_destination_name(encryptable, receiver)
     existing_names = receiver.inbox_folder.encryptables.pluck(:name)
     is_file = encryptable.is_a?(Encryptable::File)
@@ -45,8 +51,9 @@ class EncryptableTransfer
     )
   end
 
-  def new_transfer_password
-    Crypto::Symmetric::Aes256.random_key
+  def receiver_encryption_algorithm(receiver)
+    encryption_algorithm = receiver.personal_team.encryption_algorithm
+    Crypto::Symmetric::ALGORITHMS[encryption_algorithm]
   end
 
   def transfered_name(name, existing_names, is_file)