+++ frontend/app/models/team.js
@@ -4,6 +4,8 @@ export default class Team extends Model {
   @attr("string") name;
   @attr("string") description;
   @attr("string") type;
+  @attr("string") encryptionAlgorithm;
+  @attr("number") passwordBitsize;
   @attr("boolean") private;
   @attr("boolean") favourised;
   @attr("boolean") deletable;