+++ spec/unit/utils/crypto/symmetric/aes256_spec.rb
@@ -2,6 +2,7 @@
 
 require 'spec_helper'
 
+require_relative '../../../../../app/utils/crypto/symmetric'
 require_relative '../../../../../app/utils/crypto/symmetric/aes256'
 require_relative '../../../../../app/utils/crypto/rsa'